refactor: move AuthMethod and AuthHandle types to authHandle.ts
