+++ src/features/deposit/components/DepositForm/index.tsx
@@ -21,8 +21,8 @@ import { getPOABridgeInfo } from "../../../../features/machines/poaBridgeInfoAct
 import { useModalStore } from "../../../../providers/ModalStoreProvider"
 import { getAvailableDepositRoutes } from "../../../../services/depositService"
 import { ModalType } from "../../../../stores/modalStore"
+import type { AuthMethod } from "../../../../types/authHandle"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../../types/base"
-import type { AuthMethod } from "../../../../types/deposit"
 import { BlockchainEnum } from "../../../../types/interfaces"
 import type { SwappableToken } from "../../../../types/swap"
 import { isBaseToken, isUnifiedToken } from "../../../../utils/token"
