+++ src/features/gift/components/GiftHistoryWidget.tsx
@@ -2,8 +2,8 @@ import { useMemo } from "react"
 import { WidgetRoot } from "../../../components/WidgetRoot"
 import type { SignerCredentials } from "../../../core/formatters"
 import { SwapWidgetProvider } from "../../../providers/SwapWidgetProvider"
+import type { AuthMethod } from "../../../types/authHandle"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
-import type { AuthMethod } from "../../../types/deposit"
 import type { GiftLinkData } from "../types/sharedTypes"
 import { GiftHistory } from "./shared/GiftHistory"
