+++ src/features/machines/swapIntentMachine.ts
@@ -5,9 +5,9 @@ import { settings } from "../../constants/settings"
 import { logger } from "../../logger"
 import { publishIntent } from "../../services/intentService"
 import type { AggregatedQuote } from "../../services/quoteService"
+import type { AuthMethod } from "../../types/authHandle"
 import type { BaseTokenInfo, TokenValue } from "../../types/base"
 import type { Nep413DefuseMessageFor_DefuseIntents } from "../../types/defuse-contracts-types"
-import type { AuthMethod } from "../../types/deposit"
 import type { WalletMessage, WalletSignatureResult } from "../../types/swap"
 import { assert } from "../../utils/assert"
 import type { IntentsUserId } from "../../utils/defuse"
