+++ src/features/machines/withdrawUIMachine.ts
@@ -10,8 +10,8 @@ import {
 } from "xstate"
 import { logger } from "../../logger"
 import type { QuoteResult } from "../../services/quoteService"
+import type { AuthMethod } from "../../types/authHandle"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../types/base"
-import type { AuthMethod } from "../../types/deposit"
 import { assert } from "../../utils/assert"
 import { authHandleToIntentsUserId } from "../../utils/defuse"
 import {
