+++ src/features/otcDesk/components/OtcMakerForm.tsx
@@ -13,9 +13,9 @@ import { useModalController } from "../../../hooks/useModalController"
 import { useTokensUsdPrices } from "../../../hooks/useTokensUsdPrices"
 import { useTokensStore } from "../../../providers/TokensStoreProvider"
 import { ModalType } from "../../../stores/modalStore"
+import type { AuthMethod } from "../../../types/authHandle"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
 import type { MultiPayload } from "../../../types/defuse-contracts-types"
-import type { AuthMethod } from "../../../types/deposit"
 import type { RenderHostAppLink } from "../../../types/hostAppLink"
 import type { SwappableToken } from "../../../types/swap"
 import { assert } from "../../../utils/assert"
