+++ src/features/swap/components/SwapSubmitter.tsx
@@ -1,7 +1,7 @@
 import { type ReactNode, createContext } from "react"
 import { nearClient } from "../../../constants/nearClient"
 import { logger } from "../../../logger"
-import type { AuthMethod } from "../../../types/deposit"
+import type { AuthMethod } from "../../../types/authHandle"
 import { SwapUIMachineContext } from "./SwapUIMachineProvider"
 
 export const SwapSubmitterContext = createContext<{
