+++ src/services/depositService.ts
@@ -30,12 +30,9 @@ import type { depositGenerateAddressMachine } from "../features/machines/deposit
 import { getNearTxSuccessValue } from "../features/machines/getTxMachine"
 import type { storageDepositAmountMachine } from "../features/machines/storageDepositAmountMachine"
 import { logger } from "../logger"
+import { AuthMethod } from "../types/authHandle"
 import type { BaseTokenInfo, SupportedChainName } from "../types/base"
-import {
-  AuthMethod,
-  type SendTransactionEVMParams,
-  type Transaction,
-} from "../types/deposit"
+import type { SendTransactionEVMParams, Transaction } from "../types/deposit"
 import { BlockchainEnum } from "../types/interfaces"
 import { assert } from "../utils/assert"
 import { type IntentsUserId, authHandleToIntentsUserId } from "../utils/defuse"
