+++ src/types/deposit.ts
@@ -1,27 +1,7 @@
 import type { Transaction as TransactionSolana } from "@solana/web3.js"
 import type { Address, Hash } from "viem"
+import type { AuthHandle } from "./authHandle"
 import type { SwappableToken } from "./swap"
-
-export type AuthMethod = "near" | "evm" | "solana" | "webauthn"
-
-export const AuthMethod = {
-  Near: "near",
-  EVM: "evm",
-  Solana: "solana",
-  WebAuthn: "webauthn",
-} as const
-
-/**
- * Represents a public identifier used for authentication.
- * This could be a blockchain address, account name, or public key.
- * This is always a public value - never contains private keys or passwords.
- */
-export type AuthIdentifier = string
-
-export type AuthHandle = {
-  identifier: AuthIdentifier
-  method: AuthMethod
-}
 
 export type DepositWidgetProps = {
   userAddress: AuthHandle["identifier"] | undefined
